+++ CHANGES.rst
@@ -15,6 +15,7 @@ Unreleased
   :meth:`~reader._storage.Storage.delete_entries`
   storage method.
   (:issue:`301`, :issue:`96`)
+* Update vendored ``reader._http_utils`` to werkzeug 2.3.2.
 
 
 Version 3.5